Deduplicate OEmbed view
